Use safe allocation methods for Buffer instances
Also improves the way we emit `message` events.

Previously, we were binding to `data` inside the `Parser` constructor. This had the unfortunate effect of unpausing the stream
immediately, regardless of whether or not anyone was actually listening to `data` events.

In theory, this meant that message received _prior_ to additional `data`/`message` handlers being bound may be essentially lost. That is, we would emit a `data` event and clear the internal buffer without anyone ever having "heard" about it.

The new approach treats `message` the same way as `stream.Readable` treats `data`, which is that it automatically resumes the stream when a handler is bound. This means that `Parser` streams will generate backpressure now, where they may not have previously.