Support inflate/deflate via zlib
Note: I chose synchronous inflate/deflate here in order to avoid branched expectations on read/write.

Prior to the inclusion of compression, writes and reads in missive were always synchronous. By making reads and writes async for compression, we'd (potentially) make it significally more difficult for users to enable compression without the need to refactor their applications.

An example of this is creating a single, shared `encode()` stream that is used to encode messages for *all* sockets. This is way more efficient than creating an `encode()` stream per socket.

In those cases, it's convienient to be able to write to `encode()` and then synchronously `encode.read()` to retrieve the encoded buffer.

So, anyway -- the point is that I didn't want to break that with the zlib add.